Correct IPv4 MTU threshhold for clearing DF bit
The DF bit should only be cleared for generated IPv4 packets that are
smaller than or equal to 1260 bytes, as this would correspond to an IPv6
packet smaller than or equal to 1280 bytes. Therefore, PMTUD can operate
normally across the translator with IPv4 paths >1260 (IPv6 paths >1280),
so there's no need to clear the DF bit.