Remove text regarding v6 frags to v4 frags
This text is out of place, as it has nothing to do with accomodating
differences in minimum MTU betwene IPv4 and IPv6. Rather, it simply
described how the stateless translator should translate preexisting
fragments. This is however redundant, as the process is described in
RFC 6145 section 5.1.1.

Besides, the text was a bit confusing. It starts talking about "an IPv6
packet" singular, but finishes talking about "all of the IPv6
fragments" plural.