Note that protocols built on SIIT are impacted
Make it clear that it is not only SIIT that is impacted by this draft,
but that it by extension impacts protocols that build on SIIT, such as
Stateful NAT64 (and probably MAP-T, but that's not spelled out).

Alberto Leiva made the observation that this was not clear - credit him
for doing so.