Fix ambiguity in RFC6145 update
Alberto Leiva noticed that the text describing the setting of the IPv4
DF flag when translating from IPv6 contained an ambiguity (that is also
present in RFC6145): It says that the DF should be set/unset depending
on "the packet size". However it doesn't make it clear whether "the
packet" refers to the original IPv6 packet, or the translated IPv4
packet.